pyproject.toml
[project]
name = "itsfergus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "requests-aws4auth>=1.3.1",
]

[tool.ruff.lint]
extend-select = ["I"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
